Move ThemingDialogViewModel to headless Gum.Presentation (#3754)
Split IThemingService/IEffectiveThemeSettings out of ThemingDialogViewModel.cs into Gum.Presentation (both already used only System.Drawing.Color); the WPF ThemingService implementation stays in the Gum tool project. Converted the VM's 6 color properties from System.Windows.Media.Color to System.Drawing.Color per ADR-0004, dropping the now-dead WPF<->Drawing conversion helpers; the WPF View converts at the binding boundary via a new DrawingColorToMediaColorConverter. ThemeSettings/ThemeDefaultsProvider moved alongside the VM (both headless); the GeneralSettingsFile-coupled MigrateExplicitLegacyColors helper stayed behind in the Gum project as ThemeSettingsMigration since GeneralSettingsFile is WinForms-entangled.

The Microsoft.Xna.Framework.Graphics using in the original combined file was dead - dropped it, no XNA/RenderingLibrary coupling existed in this VM.

ThemingDialogView now pairs with its VM via [Dialog(typeof(ThemingDialogViewModel))] instead of same-assembly naming convention. DialogViewResolverTests' co-located-naming-convention test lost its last production example (ThemingDialogViewModel was it), so it now pins that resolver path with a synthetic same-assembly VM/View pair instead.

Part of #3754
